spotify-api/requirements.txt
Flask==2.3.3
spotipy==2.22.1
flask-cors==4.0.0
python-dotenv==1.0.0
